iam_service/src/test/java/org/path/iam/manager/NotificationStateRoleMappingManagerTest.java
package org.path.iam.manager;

import org.path.iam.dao.NotificationStateRoleMappingDao;
import org.path.iam.model.NotificationStateRoleMapping;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import java.util.Arrays;
import java.util.List;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.Mockito.when;

public class NotificationStateRoleMappingManagerTest {

    @Mock
    private NotificationStateRoleMappingDao notificationStateRoleMappingDao;

    @InjectMocks
    private NotificationStateRoleMappingManager notificationStateRoleMappingManager;

    @BeforeEach
    public void setUp() {
        MockitoAnnotations.initMocks(this);
    }

    @Test
    public void testFindByStateAndCategory() {
        // Mocking data
        String state = "example_state";
        Long categoryId = 1L;
        NotificationStateRoleMapping mapping1 = new NotificationStateRoleMapping(/* constructor arguments */);
        NotificationStateRoleMapping mapping2 = new NotificationStateRoleMapping(/* constructor arguments */);
        List<NotificationStateRoleMapping> expectedMappings = Arrays.asList(mapping1, mapping2);

        // Mocking behavior of DAO
        when(notificationStateRoleMappingDao.findByStateAndCategory(state, categoryId))
                .thenReturn(expectedMappings);

        // Calling the manager method
        List<NotificationStateRoleMapping> actualMappings = notificationStateRoleMappingManager.findByStateAndCategory(state, categoryId);

        // Verifying results
        assertEquals(expectedMappings.size(), actualMappings.size());
        // You might need to add more assertions depending on the behavior you expect
    }

    // Add more test methods as needed to cover other scenarios
}
